Use enabled false rather than hacky workaround
